requirements.txt
OpenCv-python
numpy 
ultralytics
torch
scipy
requests
matplotlib
pyyaml
cvzone
pillow
tk
pyodbc


#Extension - sqltool